+++ src/messages/att_euler.rs
@@ -1,0 +1,31 @@
+use binrw::binrw;
+
+// Attitude Euler Block 5938
+#[binrw]
+#[derive(Debug)]
+pub struct AttEuler {
+    #[br(map = |x| if x == crate::DO_NOT_USE_U4 { None } else { Some(x) })]
+    pub tow: Option<u32>,
+    #[br(map = |x| if x == crate::DO_NOT_USE_U2 { None } else { Some(x) })]
+    pub wnc: Option<u16>,
+    #[br(map = |x| if x == crate::DO_NOT_USE_U1 { None } else { Some(x) })]
+    pub nrsv: Option<u8>,
+    // TODO: create Error enum
+    pub error: u8,
+    pub mode: u16,
+    _reserved: u16,
+
+    #[br(map = |x| if x == crate::DO_NOT_USE_F4 { None } else { Some(x) })]
+    pub heading: Option<f32>,
+    #[br(map = |x| if x == crate::DO_NOT_USE_F4 { None } else { Some(x) })]
+    pub pitch: Option<f32>,
+    #[br(map = |x| if x == crate::DO_NOT_USE_F4 { None } else { Some(x) })]
+    pub roll: Option<f32>,
+
+    #[br(map = |x| if x == crate::DO_NOT_USE_F4 { None } else { Some(x) })]
+    pub pitch_dot: Option<f32>,
+    #[br(map = |x| if x == crate::DO_NOT_USE_F4 { None } else { Some(x) })]
+    pub roll_dot: Option<f32>,
+    #[br(map = |x| if x == crate::DO_NOT_USE_F4 { None } else { Some(x) })]
+    pub heading_dot: Option<f32>,
+}
